split region cache into separate files
